requirements.txt
alembic==1.15.1
annotated-types==0.7.0
anthropic==0.49.0
CherryPy==18.10.0
cherrypy-cors==1.7.0
cryptography==44.0.2
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.164.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-generativeai==0.8.4
googleapis-common-protos==1.69.1
langchain-anthropic==0.3.10
langchain-core==0.3.45
langchain-groq==0.3.0
langchain-mcp-adapters==0.0.5
langgraph==0.3.11
langgraph-checkpoint==2.0.20
langgraph-prebuilt==0.1.3
langgraph-sdk==0.1.57
langsmith==0.3.15
marshmallow==3.26.1
mcp==1.4.1
mysqlclient==2.2.7
oauthlib==3.2.2
PyJWT==2.10.1
python-dotenv==1.0.1
SQLAlchemy==2.0.39

